Properly limit string length after leading whitespace removal.
Now that we are potentially moving idx to > 0 while dealing with
whitespace we should set an upper bound on finding the end of the
scheme at len - idx, not len.

In fact, since the optimizer should have no issues figuring out that
idx is 0 in some of these cases let's just consistently apply
  start:  &arr[idx]
  length: len - idx
to everywhere.