Choose sv_catpv[n] or sv_catpv[n]_flags based on uri->is_iri
Defined two new macros, uri_sv_catpv and uri_sv_catpvn, which use
uri->is_iri to select the perlapi macro based on whether to assume the
input bytes are utf8 octets. When is_iri is non-zero, sv_catpv[n]_flags
is used with the SV_CATUTF8 flag to ensure the bytes are treated
properly when encoding the string.